refactor: 판매 관련 기능 폴더 생성, 폴더 구조 변경 및 이동
