+++ docs/codex/store/01-plan/매장_판매영역_통합_설계안.md
@@ -1,0 +1,209 @@
+# 매장 판매영역 통합 설계안
+
+## 1. 개요
+
+매장 관리자 판매 영역은 `운영형 분리`를 기준으로 설계한다.  
+사이드바는 `판매 관리`를 상위 메뉴로 두고, 그 아래에 `판매 등록`, `판매 내역`, `판매 분석` 3개 하위 메뉴를 둔다.
+
+초기 MVP는 `판매 등록`과 `판매 내역` 구현에 집중하고, `판매 분석`은 화면을 크게 만들지 않더라도 이후 함께 팔린 상품 분석이 가능하도록 판매 데이터 구조를 먼저 맞춘다.
+
+## 2. 핵심 설계 원칙
+
+- 판매는 단건 버튼형이 아니라 `판매 목록` 기반으로 처리한다.
+- 판매 데이터는 `Sale + SaleItem[]` 구조로 저장한다.
+- 재고 차감은 `SKU(상품 + 컬러 + 사이즈)` 단위로 수행한다.
+- 초기 메타데이터는 `판매시각`, `총액`, `총수량`, `판매상품 목록`, `매장`까지만 저장한다.
+- 고객정보와 결제수단은 저장하지 않는다.
+- 취소/반품은 초기 범위에서 제외한다.
+
+## 3. 내비게이션 및 용어
+
+- 매장 사이드바에 `판매 관리` 상위 메뉴를 추가한다.
+- 하위 메뉴는 `판매 등록`, `판매 내역`, `판매 분석`으로 구성한다.
+- 기존 `POS / 판매`는 `판매 관리 > 판매 등록`으로 흡수한다.
+- 사용자 화면 용어는 `장바구니`보다 `판매 목록`을 사용한다.
+- 내부 구현 상태명은 `cart`를 써도 되지만, 사용자 노출 문구는 `판매 목록`으로 통일한다.
+
+## 4. 판매 등록 화면
+
+### 4.1 레이아웃
+
+- 좌측: 상품 탐색 영역
+- 우측: 판매 목록 영역
+
+### 4.2 탐색 방식
+
+- 탐색 UX는 `카테고리 우선 + 검색 보조`로 설계한다.
+- 카테고리는 `대분류 + 소분류`를 모두 사용한다.
+
+대분류:
+
+- 상의
+- 바지
+- 치마
+- 아우터
+
+소분류 예시:
+
+- 반팔
+- 긴팔
+- 셔츠
+- 니트
+- 후드티
+- 청바지
+- 반바지
+- 긴바지
+- 츄리닝
+- 미니스커트
+- 롱스커트
+- 패딩
+- 후드집업
+- 자켓
+- 가디건
+
+### 4.3 판매 단위
+
+판매 단위는 `상품`이 아니라 `SKU`로 본다.
+
+SKU 구성 요소:
+
+- 상품명
+- `mainCategory`
+- `subCategory`
+- `color`
+- `size`
+- `unitPrice`
+- `stock`
+
+### 4.4 좌측 상품 리스트 표시 우선순위
+
+- 상품명
+- 색상 / 사이즈
+- `대분류 > 소분류`
+- 가격 / 재고
+
+### 4.5 우측 판매 목록 표시 항목
+
+- 상품명
+- 색상 / 사이즈
+- 수량
+- 라인 금액
+- 총수량 / 총금액
+- 판매 확정 버튼
+
+### 4.6 동작 규칙
+
+- 같은 SKU를 다시 선택하면 새 줄을 만들지 않고 수량만 누적한다.
+- 판매 확정 시 판매 목록 전체를 하나의 판매건으로 저장한다.
+- 저장 성공 후 모든 SKU 재고를 한 번에 차감한다.
+- 품절 SKU는 추가할 수 없다.
+- 재고 초과 수량은 추가/수정/확정 모두 차단한다.
+- 판매 완료 후 판매 목록은 초기화한다.
+
+## 5. 판매 데이터 구조
+
+판매는 반드시 `판매건 헤더 + 판매 라인` 구조로 저장한다.
+
+### 5.1 Sale
+
+- `saleId`
+- `storeId`
+- `soldAt`
+- `totalQuantity`
+- `totalAmount`
+
+### 5.2 SaleItem
+
+- `saleId`
+- `skuId`
+- `productId`
+- `productName`
+- `mainCategory`
+- `subCategory`
+- `color`
+- `size`
+- `unitPrice`
+- `quantity`
+- `lineAmount`
+
+### 5.3 저장하지 않는 정보
+
+- 고객정보
+- 회원 여부
+- 결제수단
+- 결제 승인 정보
+- 취소/반품 상태
+
+## 6. 판매 내역 화면
+
+판매 내역은 `판매건 중심`으로 조회한다.
+
+### 6.1 목록 컬럼
+
+- 판매시각
+- 판매건 번호
+- 대표 상품명
+- 총수량
+- 총금액
+
+### 6.2 표시 방식
+
+- 목록에서는 `대표 상품명 외 N건` 형태를 허용한다.
+- 카테고리는 목록에서 항상 크게 강조할 필요는 없지만, 필요 시 보조 정보로 노출할 수 있다.
+- 상세 화면에서는 각 라인에 반드시 아래 정보를 표시한다.
+
+상세 라인 정보:
+
+- 상품명
+- `대분류 > 소분류`
+- 색상
+- 사이즈
+- 수량
+- 단가
+- 금액
+
+### 6.3 상세 목적
+
+- 한 판매건에 어떤 상품들이 함께 팔렸는지 그대로 볼 수 있어야 한다.
+- 이후 연관 상품 분석의 원천 데이터가 되는 화면으로 사용한다.
+
+## 7. 판매 분석 화면
+
+초기 MVP에서는 최소 뼈대만 둔다.
+
+- 우선 목표는 분석 화면 완성이 아니라 `분석 가능한 판매 데이터 구조 확보`다.
+- 이후 확장 후보는 다음과 같다.
+
+- 상품별 판매량
+- 대분류/소분류별 매출
+- 색상/사이즈별 판매량
+- 시간대별 판매 추이
+- 함께 팔린 상품 TOP
+
+## 8. 구현 포인트
+
+- 기존 `한 줄 상품 즉시 판매 버튼` 구조는 제거하거나 보조 기능으로 축소한다.
+- 판매 확정은 반드시 `판매 목록 전체 기준`으로 처리한다.
+- 상품명만 보여주는 구조는 피하고, 최소한 보조 정보로 `대분류 > 소분류`를 유지한다.
+- 의류 상품 특성상 `색상` 필드는 이번 설계에서 반드시 포함한다.
+- 카테고리는 단일 문자열이 아니라 `mainCategory`, `subCategory`로 분리한다.
+- 고객 관리나 결제 분석 기능은 이번 범위에서 제외한다.
+
+## 9. 테스트 기준
+
+- 서로 다른 SKU 여러 개를 한 판매건에 담아 한 번에 판매할 수 있어야 한다.
+- 같은 SKU를 여러 번 추가하면 판매 목록 수량만 누적되어야 한다.
+- 판매 확정 후 모든 라인아이템 재고가 정확히 차감되어야 한다.
+- 재고보다 많은 수량은 판매 목록 추가/수정/확정 시 차단되어야 한다.
+- 품절 SKU는 판매 목록에 담을 수 없어야 한다.
+- 판매 내역 목록에는 판매건이 1건으로 보이고, 상세에서는 여러 품목 라인이 보여야 한다.
+- `saleId` 기준으로 같은 판매건의 상품 조합을 추출할 수 있어야 한다.
+- `mainCategory`, `subCategory`, `color`, `size`가 판매 데이터에 모두 남아야 한다.
+
+## 10. 가정 및 기본값
+
+- 판매는 오프라인 매장 손님 대상 현장 판매를 전제로 한다.
+- 고객정보는 재고관리 범위 밖으로 보고 저장하지 않는다.
+- 결제수단 및 결제 패턴 분석은 개발 범위 밖으로 본다.
+- 초기 구현에서는 취소/반품을 다루지 않는다.
+- 사용자 노출 용어는 `판매 목록`, 내부 데이터 구조는 `Sale`, `SaleItem`, `SKU` 기준으로 통일한다.
