+++ docs/codex/store/02-implementation/매장_판매영역_구현_내역.md
@@ -1,0 +1,247 @@
+# 매장 판매영역 구현 내역
+
+## 1. 구현 목적
+
+매장 관리자 판매 기능을 `판매 관리` 중심 구조로 재편하고, 이후 판매 분석까지 확장 가능한 데이터 구조를 선반영하는 것이 이번 구현의 목적이었다.
+
+핵심 구현 목표:
+
+- 판매 메뉴 구조 계층화
+- 의류 도메인 기반 SKU 재고 모델 도입
+- 여러 상품을 한 판매건으로 저장하는 판매 목록 흐름 구현
+- 판매 내역 조회 및 상세 확인 기능 추가
+- 판매 분석 확장을 위한 최소 화면 및 데이터 준비
+
+## 2. 반영된 구조 변경
+
+### 2.1 메뉴 구조
+
+`src/config/roleMenus.js`
+
+매장 메뉴를 다음과 같이 재구성했다.
+
+- 대시보드
+- 판매 관리
+  - 판매 등록
+  - 판매 내역
+  - 판매 분석
+- 주문 관리
+- 재고 관리
+- 입고 관리
+- AI 리포트
+
+또한 매장 로그인 후 기본 진입 경로를 `/store/sales/register`로 변경했다.
+
+### 2.2 라우터 구조
+
+`src/router/index.js`
+
+추가/변경된 라우트:
+
+- `/store/sales/register`
+- `/store/sales/history`
+- `/store/sales/analysis`
+- 기존 `/store/pos`는 `store-sales-register`로 리다이렉트
+
+## 3. 데이터 모델 구현
+
+### 3.1 재고 모델
+
+`src/stores/inventory.js`
+
+기존 공산품 중심 더미 재고를 제거하고, 의류 도메인 기반 `SKU` 재고 모델로 교체했다.
+
+반영 내용:
+
+- `mainCategory`
+- `subCategory`
+- `color`
+- `size`
+- `unitPrice`
+- `stock`
+- `safetyStock`
+
+지원 기능:
+
+- 대분류 목록 계산
+- 대분류별 소분류 목록 계산
+- 색상 옵션 목록 계산
+- SKU 상태 판정(`정상`, `부족`, `품절`)
+- 다중 SKU 판매 차감(`sellItems`)
+
+### 3.2 판매 모델
+
+`src/stores/sales.js`
+
+새 판매 스토어를 추가했다.
+
+구현 내용:
+
+- `Sale` 저장
+- `SaleItem[]` 저장
+- 판매번호(`saleId`) 자동 생성
+- 판매 시각 저장
+- 총수량 / 총금액 계산
+- localStorage 영속화
+
+저장 필드:
+
+- `saleId`
+- `storeId`
+- `soldAt`
+- `totalQuantity`
+- `totalAmount`
+- `items[]`
+
+`items[]` 필드:
+
+- `skuId`
+- `productId`
+- `productName`
+- `mainCategory`
+- `subCategory`
+- `color`
+- `size`
+- `unitPrice`
+- `quantity`
+- `lineAmount`
+
+## 4. 화면 구현 내역
+
+### 4.1 판매 등록
+
+`src/views/store/StorePosView.vue`
+
+기존 한 줄 즉시 판매 UI를 제거하고, `좌측 상품 탐색 + 우측 판매 목록` 구조로 전면 개편했다.
+
+반영 기능:
+
+- 대분류 필터
+- 소분류 필터
+- 색상 필터
+- 검색 필터
+- SKU 목록 조회
+- 품절/부족/정상 상태 표시
+- 판매 목록 추가
+- 같은 SKU 재추가 시 수량 누적
+- 판매 목록 수량 직접 수정
+- 판매 목록 개별 삭제
+- 판매 목록 전체 비우기
+- 총수량/총금액 계산
+- 판매 확정 시 판매건 저장 + SKU 재고 차감
+
+사용자 노출 용어:
+
+- `장바구니` 대신 `판매 목록`
+
+### 4.2 판매 내역
+
+`src/views/store/StoreSalesHistoryView.vue`
+
+판매건 중심 조회 화면을 추가했다.
+
+반영 기능:
+
+- 판매건 목록 조회
+- 판매번호 / 상품명 / 카테고리 검색
+- 판매건 선택
+- 판매 상세 조회
+- 함께 팔린 상품 구성 확인
+
+목록 컬럼:
+
+- 판매시각
+- 판매번호
+- 대표 상품명
+- 총수량
+- 총금액
+
+상세 컬럼:
+
+- 상품명
+- `대분류 > 소분류`
+- 색상 / 사이즈
+- 수량
+- 단가
+- 금액
+
+### 4.3 판매 분석
+
+`src/views/store/StoreSalesAnalysisView.vue`
+
+초기 MVP 범위에서는 최소 뼈대만 구현했다.
+
+반영 기능:
+
+- 판매건 수
+- 누적 판매 수량
+- 누적 판매 금액
+- 상품별 누적 판매 미리보기
+
+이 화면은 현재 완성형 분석이 아니라, 이후 분석 기능 확장을 위한 진입점이다.
+
+## 5. 주변 화면 정리
+
+판매 관리가 상위 메뉴가 되면서, 다른 매장 화면도 상단/사이드 메뉴 연결 방식을 정리했다.
+
+수정 대상:
+
+- `src/views/store/StoreDashboardView.vue`
+- `src/views/store/StoreOrdersView.vue`
+- `src/views/store/StoreInventoryView.vue`
+- `src/views/store/StoreInboundView.vue`
+- `src/views/store/StoreAiReportView.vue`
+
+정리 내용:
+
+- 각 화면이 `roleMenus.store` 전체를 top menu로 사용
+- 각 화면이 자기 메뉴의 children만 side menu로 사용
+- 대시보드의 POS 바로가기를 `/store/sales/register`로 변경
+
+## 6. 구현 시 제외한 범위
+
+이번 구현에서는 아래 항목을 의도적으로 제외했다.
+
+- 고객정보 저장
+- 회원/비회원 구분
+- 결제수단 저장
+- 결제 패턴 분석
+- 취소/반품 처리
+- 고급 판매 분석(시간대별, 연관상품 TOP, 카테고리별 매출 시각화)
+
+## 7. 검증 내역
+
+실행 검증:
+
+- `npm.cmd exec vite build -- --configLoader native`
+
+검증 결과:
+
+- 프로덕션 빌드 성공
+
+비고:
+
+- 번들 크기 경고는 있었지만 빌드 자체는 정상 완료
+- `git status`는 현재 실행 환경의 `safe.directory` 제한으로 확인하지 못함
+
+## 8. 다음 단계 제안
+
+다음 작업 후보:
+
+- 판매 내역 기간 필터 추가
+- 시간대별 판매 집계 추가
+- 함께 팔린 상품 조합 분석 추가
+- 카테고리별 판매량/매출 분석 추가
+- 취소/반품 도입 전 상태 모델 확장
+
+## 9. 관련 파일
+
+이번 구현에 직접적으로 관련된 주요 파일:
+
+- `src/config/roleMenus.js`
+- `src/router/index.js`
+- `src/stores/inventory.js`
+- `src/stores/sales.js`
+- `src/views/store/StorePosView.vue`
+- `src/views/store/StoreSalesHistoryView.vue`
+- `src/views/store/StoreSalesAnalysisView.vue`
