+++ src/router/index.js
@@ -7,9 +7,9 @@ import LoginView from '@/views/LoginView.vue'
 import SignupView from '@/views/SignupView.vue'
 
 import StoreDashboardView from '@/views/store/StoreDashboardView.vue'
-import StorePosView from '@/views/store/StorePosView.vue'
-import StoreSalesHistoryView from '@/views/store/StoreSalesHistoryView.vue'
-import StoreSalesAnalysisView from '@/views/store/StoreSalesAnalysisView.vue'
+import StorePosView from '@/views/store/sales/StorePosView.vue'
+import StoreSalesHistoryView from '@/views/store/sales/StoreSalesHistoryView.vue'
+import StoreSalesAnalysisView from '@/views/store/sales/StoreSalesAnalysisView.vue'
 import StoreOrdersView from '@/views/store/StoreOrdersView.vue'
 import StoreInventoryView from '@/views/store/StoreInventoryView.vue'
 import StoreInboundView from '@/views/store/StoreInboundView.vue'
